dom0/domd: added recipes for Dom0 and DomD
Added the base recipes to build Dom0 and DomD for Gen5 VDK
platform based on the kirkstone meta-xt-prod-devel-rcar-gen4
and VirtiIO-related part of meta-xt-prod-devel-rcar.
Currently, Renesas Yocto BSP doesn't have a separate target
for Gen5, so Spider machine is used as a base target platform.
This commit introduces recipes for building binaries that can
be used to run U-Boot, Xen, Dom0, and DomD on VDK 3.5.1 release.
The following binaries are based on Renesas and EPAM repositories
with additional fixes and customizations:

U-Boot 2022.01:
	https://github.com/renesas-rcar/u-boot/tree/v2022.01/rcar-6.0.0.rc6_vpf.rc4
Dom0 Linux 5.10.41:
	https://github.com/renesas-rcar/linux-bsp/tree/v5.10.41/rcar-5.1.7.rc5
DomD Linux 5.10.147:
	https://github.com/renesas-rcar/linux-bsp/tree/v5.10.147/rcar-5.2.0.rc11_vpf.rc8
Xen 4.19:
	https://github.com/xen-troops/xen/tree/xen-4.19-xt0.1
QEMU for VirtIO 7.0:
	https://github.com/xen-troops/qemu/tree/v7.0.0-xt

Signed-off-by: Leonid Komarianskyi <leonid_komarianskyi@epam.com>