Reorganize project
This commit addresses the issue that makes it impossible
to properly release the product.
The root cause of the issue is the self-reference inside
yaml configuration, which requires to use of a hash/tag
that doesn't exist before the commit is merged.
The only way to solve this - is to remove self-reference,
and to consider the yaml configuration as the part of
the product, not as a separate entity.

So, this commit:
- removes self-reference from the yaml;
- moves yocto layers into the dedicated directory;
- adds .gitignore with the list of build artifacts;
- describe new build steps in the REAdME.md;
- merge DomA-related steps into the main build instruction.

Signed-off-by: Ruslan Shymkevych <ruslan_shymkevych@epam.com>
Reviewed-by: Leonid Komarianskyi <leonid_komarianskyi@epam.com>
Reviewed-by: Dmytro Firsov <dmytro_firsov@epam.com>