yaml: move common_data inside build directory
In current version of prod-devel-rcar-gen5-vdk.yaml common files for
Yocto builds are stored outside of build directory and can be shared
between different builds (if placed in correct directory level). It
speeds up build, but any build in such case can not be fully "clean".
Also it affects docker build and quickly consumes all storage in
default (not shared as "product" dir) docker volume. This feature is
mainly suitable for developers, so should not be present in main
branch, that is used for releases.

Thus, move common data inside the build directory and make it unique
for product build.

Signed-off-by: Dmytro Firsov <dmytro_firsov@epam.com>
Reviewed-by: Leonid Komarianskyi <leonid_komarianskyi@epam.com>