yaml: disable resizing of the android's images
With the moulin 0.27 the handling of the option `size` for the
`raw_image` was made compliant with the same for other types
of the images.
This means that if the `size` is specified for the `raw_image`
and the real image size is less, then we are trying to resize
the real file size to the `size`. This approach works for ext4
images. But we found that `raw_image` with `size` is used for
the android images. And mentioned change breaks the assembling
of the android.img.

This commit explicitly disables resizing for the android
images.

And `min_ver` is set to 0.27 due to usage of the `resize` option.

Signed-off-by: Ruslan Shymkevych <ruslan_shymkevych@epam.com>
Signed-off-by: Andrii Chepurnyi <andrii_chepurnyi@epam.com>
Acked-by: Leonid Komarianskyi <leonid_komarianskyi@epam.com>
Acked-by: Dmytro Firsov <dmytro_firsov@epam.com>