domd: fix using config and update it for 6.1 kernel
Due to the meta-renesas layers using '--alldefconfig' for
KCONFIG_MODE, the defconfig provided in files is ignored
despite clearing the KBUILD_DEFCONFIG variable. This
happens because while using this option, the config is
taken from kernel sources. To use custom .config file
which named 'defconfig', the KCONFIG_MODE should be set
to '--allnoconfig' option and KBUILD_DEFCONFIG must be empty.
Additionally, the config file has been updated because
of switching kernel version from 5.10 to 6.1.

Signed-off-by: Leonid Komarianskyi <leonid_komarianskyi@epam.com>
Acked-by: Dmytro Firsov <dmytro_firsov@epam.com>