domd: align dts and boot artifacts archive names
Commonly, the device tree names for domains in the product
and boot artifacts archive contain lower-case SoC names.
Previously, DomD contained an uppercase version of the SoC
name, which is inconsistent with the naming used for other
components in gen5-prod-devel and other prod-devel projects.
This update aligns the DomD DTS names and boot artifacts
archive name with the common lowercase approach.

Signed-off-by: Leonid Komarianskyi <leonid_komarianskyi@epam.com>
Acked-by: Dmytro Firsov <dmytro_firsov@epam.com>