yaml: add generation of eMMC flash data image
eMMC with SMMU functionality is confirmed since VDK 3.12.0 and it was
successfully integrated as a DomD device. To use it as a boot storage
we need to have image, that will store partition table and DomD/DomU
root fileststems as a partitions. It is required to get rid of blkback
in Dom0 and comply with thin Dom0 concept.

Add reqired mmc_flash.img to product yaml file as a first part of
migration from Virtualizer virtio.img to eMMC boot.

Signed-off-by: Dmytro Firsov <dmytro_firsov@epam.com>
Reviewed-by: Leonid Komarianskyi <leonid_komarianskyi@epam.com>