domd: reorganize Xen and DomD DTS
Commonly, we have a generic native platform DTS, which
can be included in DomD or Xen DTS and overrided. For Gen5
VDK there is no appropriate DTS. Thus, initially, we created
separate 2 DTS for Xen and DomD, but now they contain a lot
of common parts which leads to node duplication. To avoid this,
it is better to create a parent platform DTS, include it in
DomD and Xen DTS, and modify or add the required parts.

Signed-off-by: Leonid Komarianskyi <leonid_komarianskyi@epam.com>
Reviewed-by: Dmytro Firsov <dmytro_firsov@epam.com>